requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
face-recognition>=1.3.0
Pillow>=8.0.0
scipy>=1.7.0
watchdog>=2.1.0
tqdm>=4.62.0
imagehash>=4.2.0
